+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/DefaultAudioVideoFacade.swift
@@ -41,14 +41,18 @@ import Foundation
         self.eventAnalyticsController = eventAnalyticsController
     }
 
-    public func start(callKitEnabled: Bool = false) throws {
-        try audioVideoController.start(callKitEnabled: callKitEnabled)
+    public func start(audioVideoConfiguration: AudioVideoConfiguration) throws {
+        try audioVideoController.start(audioVideoConfiguration: audioVideoConfiguration)
 
-        trace(name: "start(callKitEnabled: \(callKitEnabled))")
+        trace(name: "start(audioVideoConfiguration: \(audioVideoConfiguration))")
+    }
+
+    public func start(callKitEnabled: Bool = false) throws {
+        try self.start(audioVideoConfiguration: AudioVideoConfiguration(callKitEnabled: callKitEnabled))
     }
 
     public func start() throws {
-        try self.start(callKitEnabled: false)
+        try self.start(audioVideoConfiguration: AudioVideoConfiguration())
     }
 
     public func stop() {