+++ AmazonChimeSDK/AmazonChimeSDK/internal/AttendeeStatus.swift
@@ -19,6 +19,9 @@ import Foundation
     /// The attendee dropped due to network issues
     case dropped = 3
 
+    /// The attendee joined without audio
+    case joinedNoAudio = 4
+
     public var description: String {
         switch self {
         case .joined:
@@ -27,6 +30,8 @@ import Foundation
             return "left"
         case .dropped:
             return "dropped"
+        case .joinedNoAudio:
+            return "joinedNoAudio"
         }
     }
 }