+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/DeviceSelectionViewController.swift
@@ -10,6 +10,7 @@ import AmazonChimeSDK
 import UIKit
 
 class DeviceSelectionViewController: UIViewController {
+    @IBOutlet var audioDeviceLabel: UILabel!
     @IBOutlet var audioDevicePicker: UIPickerView!
     @IBOutlet var videoDevicePicker: UIPickerView!
     @IBOutlet var videoFormatPicker: UIPickerView!
@@ -27,6 +28,10 @@ class DeviceSelectionViewController: UIViewController {
         videoDevicePicker.dataSource = self
         videoFormatPicker.delegate = self
         videoFormatPicker.dataSource = self
+
+        /// Update the visibility of Audio Device selection based on the audio mode
+        updateAudioState(audioEnabled: model?.audioVideoConfig.audioMode != .noAudio)
+
         videoPreviewImageView.mirror = model?.shouldMirrorPreview ?? false
         model?.cameraCaptureSource.addVideoSink(sink: videoPreviewImageView)
         model?.cameraCaptureSource.start()
@@ -39,6 +44,11 @@ class DeviceSelectionViewController: UIViewController {
         model.cameraCaptureSource.stop()
         model.cameraCaptureSource.removeVideoSink(sink: videoPreviewImageView)
         MeetingModule.shared().deviceSelected(model)
+    }
+
+    private func updateAudioState(audioEnabled: Bool) {
+        audioDeviceLabel.isHidden = !audioEnabled
+        audioDevicePicker.isHidden = !audioEnabled
     }
 }