+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/JoiningViewController.swift
@@ -15,10 +15,12 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
     @IBOutlet var meetingIdTextField: UITextField!
     @IBOutlet var nameTextField: UITextField!
     @IBOutlet var versionLabel: UILabel!
-    @IBOutlet var joinWithoutCallKitButton: UIButton!
-    @IBOutlet var joinAsIncomingCallButton: UIButton!
-    @IBOutlet var joinAsOutgoingCallButton: UIButton!
+    @IBOutlet var callKitOptionsPicker: UIPickerView!
+    @IBOutlet var audioSwitch: UISwitch!
+    @IBOutlet var joinButton: UIButton!
     @IBOutlet var debugSettingsButton: UIButton!
+
+    var callKitOptions = ["Don't use CallKit", "CallKit as Incoming in 10s", "CallKit as Outgoing"]
 
     private let toastDisplayDuration = 2.0
     private let mainStoryboard = UIStoryboard(name: "Main", bundle: nil)
@@ -29,26 +31,41 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
         meetingIdTextField.delegate = self
         nameTextField.delegate = self
 
+        callKitOptionsPicker.delegate = self
+        callKitOptionsPicker.dataSource = self
+
         setupHideKeyboardOnTap()
         versionLabel.text = "amazon-chime-sdk-ios@\(Versioning.sdkVersion())"
     }
 
     override func viewWillAppear(_: Bool) {
-        setJoinButtons(isEnabled: true)
+        callKitOptionsPicker.selectRow(0, inComponent: 0, animated: false)
+        joinButton.isEnabled = true
     }
 
-    @IBAction func joinWithoutCallKitButtonClicked(_: UIButton) {
-        joinMeeting(callKitOption: .disabled)
-    }
+    @IBAction func joinButton(_: UIButton) {
+        // CallKit Option
+        var callKitOption: CallKitOption = .disabled
+        switch callKitOptionsPicker.selectedRow(inComponent: 0) {
+        case 1:
+            callKitOption = .incoming
+            view.makeToast("You can background the app or lock screen while waiting",
+                           duration: incomingCallKitDelayInSeconds)
+        case 2:
+            callKitOption = .outgoing
+        default:
+            callKitOption = .disabled
+        }
 
-    @IBAction func joinAsIncomingCallButton(_: UIButton) {
-        view.makeToast("You can background the app or lock screen while waiting",
-                       duration: incomingCallKitDelayInSeconds)
-        joinMeeting(callKitOption: .incoming)
-    }
+        // Audio Mode
+        var audioMode: AudioMode = .mono
+        if !audioSwitch.isOn {
+            audioMode = .noAudio
+        }
 
-    @IBAction func joinAsOutgoingCallButton(_: UIButton) {
-        joinMeeting(callKitOption: .outgoing)
+        joinMeeting(audioVideoConfig: AudioVideoConfiguration(audioMode: audioMode, callKitEnabled: callKitOption != .disabled),
+                    callKitOption: callKitOption
+        )
     }
 
     @IBAction func debugSettingsButtonClicked (_: UIButton) {
@@ -66,13 +83,7 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
         return true
     }
 
-    private func setJoinButtons(isEnabled: Bool) {
-        joinWithoutCallKitButton.isEnabled = isEnabled
-        joinAsIncomingCallButton.isEnabled = isEnabled
-        joinAsOutgoingCallButton.isEnabled = isEnabled
-    }
-
-    func joinMeeting(callKitOption: CallKitOption) {
+    func joinMeeting(audioVideoConfig: AudioVideoConfiguration, callKitOption: CallKitOption) {
         view.endEditing(true)
         let meetingId = meetingIdTextField.text ?? ""
         let name = nameTextField.text ?? ""
@@ -87,6 +98,7 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
 
         MeetingModule.shared().prepareMeeting(meetingId: meetingId,
                                               selfName: name,
+                                              audioVideoConfig: audioVideoConfig,
                                               option: callKitOption,
                                               overriddenEndpoint: debugSettingsModel.endpointUrl) { success in
             DispatchQueue.main.async {
@@ -97,5 +109,24 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
                 }
             }
         }
+    }
+}
+
+extension JoiningViewController: UIPickerViewDelegate {
+    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent _: Int) -> String? {
+        if row >= callKitOptions.count {
+            return nil
+        }
+        return callKitOptions[row]
+    }
+}
+
+extension JoiningViewController: UIPickerViewDataSource {
+    func numberOfComponents(in _: UIPickerView) -> Int {
+        return 1
+    }
+
+    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent _: Int) -> Int {
+        return callKitOptions.count
     }
 }