+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/MeetingModule.swift
@@ -31,6 +31,7 @@ class MeetingModule {
 
     func prepareMeeting(meetingId: String,
                         selfName: String,
+                        audioVideoConfig: AudioVideoConfiguration,
                         option: CallKitOption,
                         overriddenEndpoint: String,
                         completion: @escaping (Bool) -> Void) {
@@ -51,6 +52,7 @@ class MeetingModule {
                 let meetingModel = MeetingModel(meetingSessionConfig: meetingSessionConfig,
                                                 meetingId: meetingId,
                                                 selfName: selfName,
+                                                audioVideoConfig: audioVideoConfig,
                                                 callKitOption: option)
                 self.meetings[meetingModel.uuid] = meetingModel